[Bugfix] Stepper: allow delete input to empty
When the stepper has value '1' and the user try to edit the value
directly to '50', the input field stop user from delete '1' to ''.

The reason for the bug is `props: { value: Number }` trying to
convert empty string to 0, then `watch: { value() {} }` will set 0
to 1 automatically.

This fix try to:

* disable type check for `value` prop

The bug was introduced from 696ae3c.

Resolves: #1151